[level 0] Title: 두 수의 나눗셈, Time: 0.00 ms, Memory: 9.25 MB -BaekjoonHub

+++ Python3/프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.py
@@ -1,2 +1,2 @@
 def solution(num1, num2):
-    return num1 / num2 * 1000 // 1
+    return int(num1 / num2 * 1000)
